fwk: rename pkg ./testdata to ./internal/fwktest
mkdir -p fwk/internal/fwktest
git mv fwk/testdata/*.go fwk/internal/fwktest
( cd fwk/internal/fwktest; sed -i '' 's/package testdata/package fwktest/' *.go )
git add fwk/internal/fwktest
