+++ fwk/internal/fwktest/inputstream.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package testdata
+package fwktest
 
 import (
 	"fmt"
@@ -11,28 +11,28 @@ import (
 	"go-hep.org/x/hep/fwk"
 )
 
-type OutputStream struct {
-	input string
-
-	W io.Writer
+type InputStream struct {
+	output string
+	R      io.Reader
 }
 
-func (out *OutputStream) Connect(ports []fwk.Port) error {
+func (stream *InputStream) Connect(ports []fwk.Port) error {
 	var err error
-	out.input = ports[0].Name
+	stream.output = ports[0].Name
+
 	return err
 }
 
-func (out *OutputStream) Write(ctx fwk.Context) error {
+func (stream *InputStream) Read(ctx fwk.Context) error {
 	var err error
 	store := ctx.Store()
-	v, err := store.Get(out.input)
+	var data int64
+	_, err = fmt.Fscanf(stream.R, "%d\n", &data)
 	if err != nil {
 		return err
 	}
 
-	data := v.(int64)
-	_, err = out.W.Write([]byte(fmt.Sprintf("%d\n", data)))
+	err = store.Put(stream.output, data)
 	if err != nil {
 		return err
 	}
@@ -40,8 +40,7 @@ func (out *OutputStream) Write(ctx fwk.Context) error {
 	return err
 }
 
-func (out *OutputStream) Disconnect() error {
+func (stream *InputStream) Disconnect() error {
 	var err error
-
 	return err
 }
